+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 - Made RadioButtonCollection and Select countable.
+- Made RadioButtonCollection iterable.
 
 ## [2.0.0] - 2018-05-14
 - Updated PHP requirement to 7.1.